# Use boost::flat_map and optimize getDelta/applyDelta