--- a/flow/flat_buffers.h
+++ b/flow/flat_buffers.h
@@ -21,6 +21,7 @@
 #pragma once
 
 #include <algorithm>
+#include <boost/container/flat_map.hpp>
 #include <iterator>
 #include <cstring>
 #include <functional>
@@ -206,6 +207,30 @@ struct vector_like_traits<std::map<Key, T, Compare, Allocator>> : std::true_type
 template <class Key, class T, class Hash, class Pred, class Allocator>
 struct vector_like_traits<std::unordered_map<Key, T, Hash, Pred, Allocator>> : std::true_type {
 	using Vec = std::unordered_map<Key, T, Hash, Pred, Allocator>;
+	using value_type = std::pair<Key, T>;
+	using iterator = typename Vec::const_iterator;
+	using insert_iterator = std::insert_iterator<Vec>;
+
+	template <class Context>
+	static size_t num_entries(const Vec& v, Context&) {
+		return v.size();
+	}
+	template <class Context>
+	static void reserve(Vec& v, size_t size, Context&) {}
+
+	template <class Context>
+	static insert_iterator insert(Vec& v, Context&) {
+		return std::inserter(v, v.end());
+	}
+	template <class Context>
+	static iterator begin(const Vec& v, Context&) {
+		return v.begin();
+	}
+};
+
+template <class Key, class T, class Compare, class Allocator>
+struct vector_like_traits<boost::container::flat_map<Key, T, Compare, Allocator>> : std::true_type {
+	using Vec = boost::container::flat_map<Key, T, Compare, Allocator>;
 	using value_type = std::pair<Key, T>;
 	using iterator = typename Vec::const_iterator;
 	using insert_iterator = std::insert_iterator<Vec>;