Introduce setVersionNoCheck() to avoid building a temporary map.
`std::map` is faster than `std::unordered_map` in benchmarks, while `boost::flat_map` beats both.

boost::flat_map is the fastest
bench_vv_getdelta/512/1024     8768606 ns      8768555 ns           80 Tags=512 getDeltaTimes=1024 items_per_second=116.781k/s
bench_vv_getdelta/1024/1024   12051740 ns     12051549 ns           58 Tags=1024 getDeltaTimes=1024 items_per_second=84.9683k/s

std::unordered_map
bench_vv_getdelta/512/1024    43800487 ns     43799915 ns           16 Tags=512 getDeltaTimes=1024 items_per_second=23.379k/s
bench_vv_getdelta/1024/1024   64894719 ns     64893852 ns           11 Tags=1024 getDeltaTimes=1024 items_per_second=15.7796k/s

std::map
bench_vv_getdelta/512/1024    34068779 ns     34068233 ns           21 Tags=512 getDeltaTimes=1024 items_per_second=30.0573k/s
bench_vv_getdelta/1024/1024   49867913 ns     49852281 ns           14 Tags=1024 getDeltaTimes=1024 items_per_second=20.5407k/s

Correctness found 2 timeout failures and 1 assertion failure (NativeAPI.actor.cpp 230), which reproduced at fe0c13797, i.e., without this PR:
```
-r simulation --crash --logsize 1024MB -f ./foundationdb/tests/rare/SwizzledLargeApiCorrectness.toml -s 220943799 -b on
-r simulation --crash --logsize 1024MB -f ./foundationdb/tests/fast/KillRegionCycle.toml -s 343565765 -b on
-r simulation --crash --logsize 1024MB -f ./foundationdb/tests/slow/DDBalanceAndRemoveStatus.toml -s 559021579 -b on
```
So these failures are unrelated to this PR.

# Code-Reviewer Section

The general guidelines can be found [here](https://github.com/apple/foundationdb/wiki/FoundationDB-Commit-Process).

Please check each of the following things and check *all* boxes before accepting a PR.

- [ ] The PR has a description, explaining both the problem and the solution.
- [ ] The description mentions which forms of testing were done and the testing seems reasonable.
- [ ] Every function/class/actor that was touched is reasonably well documented.

## For Release-Branches

If this PR is made against a release-branch, please also check the following:

- [ ] This change/bugfix is a cherry-pick from the next younger branch (younger `release-branch` or `master` if this is the youngest branch)
- [ ] There is a good reason why this PR needs to go into a release branch and this reason is documented (either in the description above or in a linked GitHub issue)